refactor: extract batch tests
